% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.load.eddy4R.R
\name{def.load.eddy4R}
\alias{def.load.eddy4R}
\title{Dummy function to ensure eddy4R.base & eddy4R.qaqc are captured in the renv lockfile}
\usage{
def.load.eddy4R()
}
\value{
eddy4R.base and eddy4R.qaqc are loaded
}
\description{
Definition function. Load eddy4R.base and eddy4R.qaqc to ensure that these packages are
documented in the renv lockfile.
}
\examples{
Currently none
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
