% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.read.thsh.qaqc.df.R
\name{def.read.thsh.qaqc.df}
\alias{def.read.thsh.qaqc.df}
\title{Read QAQC thresholds for NEON instrumented systems data products from JSON file to data frame}
\usage{
def.read.thsh.qaqc.df(NameFile = NULL, strJson = NULL, listThsh = NULL)
}
\arguments{
\item{NameFile}{Filename (including relative or absolute path). Must be json format.}

\item{strJson}{character string of data in JSON format (as produced by rjson::toJSON()). Note that
only one of NameFile, strJson, or thsh may be entered. If more than one are supplied, the first
valid input will be used.}

\item{listThsh}{A list of thresholds already read in (as produced by rjson::fromJSON()) and extracted 
from the main 'thresholds' list. Note that only one of NameFile, strJson, or listThsh may be entered. 
If more than one are supplied, the first valid input will be used.}
}
\value{
A data frame with all thresholds contained in the json file. Note that the context entries
for each threshold have been combined into a single pipe-delimited string.
}
\description{
Definition function. Read QA/QC threshold information from json file and convert to data frame.
}
\examples{
Currently none
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[NEONprocIS.qaqc]{def.read.thsh.qaqc.list}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
