% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.read.thsh.qaqc.list.R
\name{def.read.thsh.qaqc.list}
\alias{def.read.thsh.qaqc.list}
\title{Read QAQC thresholds for NEON instrumented systems data products from JSON file to list}
\usage{
def.read.thsh.qaqc.list(
  NameFile = NULL,
  strJson = NULL,
  listThsh = NULL,
  log = NULL
)
}
\arguments{
\item{NameFile}{Character value. Filename (including relative or absolute path) of the 
thresholds file. Must be json format.}

\item{strJson}{character string of data in JSON format (as produced by rjson::toJSON()). Note that
only one of NameFile, strJson, or thsh may be entered. If more than one are supplied, the first
valid input will be used.}

\item{listThsh}{A list of thresholds already read in (as produced by rjson::fromJSON() and extracted 
from the main 'thresholds' list. Note that only one of NameFile, strJson, or listThsh may be entered. 
If more than one are supplied, the first valid input will be used.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output in addition to standard R error messaging. Defaults to NULL, in which the logger will be
created and used within the function.}
}
\value{
A list of thresholds, with dates in fields start_date and end_date converted to POSIXct.
}
\description{
Definition function. Read QA/QC threshold information from json file to list. Convert timestamps to POSIX.
}
\examples{
Currently none
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\link[NEONprocIS.qaqc]{def.read.thsh.qaqc.df}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
