% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spk.mad.R
\name{def.spk.mad}
\alias{def.spk.mad}
\title{Perform spike test using the median absolute deviation method}
\usage{
def.spk.mad(
  data,
  Meth = c("A", "B")[1],
  ThshMad = 7,
  Wndw,
  WndwStep = 1,
  WndwFracSpkMin = 0.1,
  NumGrp = NULL,
  NaFracMax = 0.1,
  log = NULL
)
}
\arguments{
\item{data}{a numeric vector of observations}

\item{Meth}{Character value of either 'A' or 'B'. Method A evaluates the central point in 
each window as the window slides along the timeseries. Method B evaluates all points within
each window as it slides along the timeseries by WndwStep each time, thus evaluating a point 
multiple times. For this method, a point must be identified as a spike at least the fraction 
of the time specified in WndwFracSpkMin to be flagged as a spike. Default is A.}

\item{ThshMad}{The threshold multiple of the median absolute deviation of the median (MAD) for 
each window outside which a value must lie to be identified as a spike. Default is 7.}

\item{Wndw}{Integer window size for calculating MAD. Must be odd. If even, 1 will be added. 
If adding 1 makes Wndw longer than the length of the data, 1 will be subtracted.}

\item{WndwStep}{Method B only. An integer value specifying the number of data points to slide 
the window as it moves along the timeseries. Default is 1. Maximum is 1/2 the window size.}

\item{WndwFracSpkMin}{Method B only. Minimum threshold fraction of windows a point must fail 
the spike test to be flagged as a spike.}

\item{NumGrp}{Either NULL or an integer value specifying the minimum number of consective spikes,
 at and above which the points will be deemed plausible values and not flagged. 
Defaults to NULL, which will result in no evaluation of consecutive spikes. Default is NULL.}

\item{NaFracMax}{Fraction between 0 and 1 specifying the maximum fraction of missing values 
for which each window retains a sufficient sample size to determine the median absolute deviation.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
An integer vector the same length as data, with quality flag results of the spike test 
(0 = pass, 1 = fail, -1 = cannot eval)
}
\description{
Definition function. Evaluate the spike test - whether an observation lies outside the median 
plus/minus a multiple of the median absolute deviation computed over a surrounding window of 
observations.
}
\examples{
data <- c(1,2,3,4,3,2,1,2,3,4,50,3,2,1,2,3,4,3,2,1)
# Method A
qfSpk <- def.spk.mad(data=data,Meth='A',ThshMad=7,Wndw=5)

# Method B
qfSpk <- def.spk.mad(data=data,Meth='B',ThshMad=7,Wndw=5,WndwStep=3,WndwFracSpkMin=0.5)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000783 - ATBD: Time Series Automatic Despiking for TIS Level 1 Data Products
}
\seealso{
\link[eddy4R.qaqc]{def.dspk.wndw}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{QA/QC,}
\keyword{despike,}
\keyword{plausibility,}
\keyword{quality}
