% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.time.heat.on.R
\name{def.time.heat.on}
\alias{def.time.heat.on}
\title{Determine discrete periods a heater was turned on based on heater events}
\usage{
def.time.heat.on(dataHeat, TimeOffAuto = NULL, log = NULL)
}
\arguments{
\item{dataHeat}{Data frame of heater event data as returned by NEONprocIS.base::def.read.evnt.json.R
At a minimum, column variables include:
timestamp = POSIX time of heater status
status = logical (TRUE=heater on)}

\item{TimeOffAuto}{A difftime object indicating the timeout period after which to assume the heater
turned off even though there is no even indicating so (e.g. base::as.difftime(30,units='mins')).
Default is never (NULL).}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log 
output in addition to standard R error messaging. Defaults to NULL, in which no logger other than 
standard R error messaging will be used.}
}
\value{
A data frame of:
timeOn POSIXct Time heater turned on
timeOff POSIXct Time heater turned off
}
\description{
Definition function. Determine discrete periods a heater was turned on based on heater events.
}
\examples{
TimeOffAuto <- base::as.difftime(30,units="mins")
timeOnHeat <- def.heat.time.on(dataHeat,TimeOffAuto=TimeOffAuto)
 
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\code{\link[NEONprocIS.base]{def.log.init}}

\code{\link[NEONprocIS.base]{def.read.evnt.json}}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
